Conflicts:
	fannie/item/modules/BaseItemModule.php
	fannie/item/modules/ScaleItemModule.php
	fannie/modules/plugins2.0/HouseCoupon/HouseCouponEditor.php
	fannie/modules/plugins2.0/HouseCoupon/index.php
	pos/is4c-nf/graphics/is4c.gif
	pos/is4c-nf/gui-modules/nslogin.php
	pos/is4c-nf/install/extra_config.php
	pos/is4c-nf/install/index.php
	pos/is4c-nf/lib/AutoLoader.php
	pos/is4c-nf/lib/Database.php
	pos/is4c-nf/lib/PrehLib.php
	pos/is4c-nf/lib/ReceiptLib.php
	pos/is4c-nf/lib/SQLManager.php
	pos/is4c-nf/lib/Scanning/SpecialUPCs/HouseCoupon.php
	pos/is4c-nf/lib/TotalActions/ArOverpayAction.php
	pos/is4c-nf/lib/TotalActions/AutoCoupon.php
	pos/is4c-nf/plugins/QuickMenus/quickmenus/1.php
	pos/is4c-nf/plugins/QuickMenus/quickmenus/10.php

+++ fannie/item/laneUpdates.php
@@ -41,6 +41,11 @@ function addProductAllLanes($upc){
 		$laneupdate_sql->add_connection($FANNIE_LANES[$i]['host'],$FANNIE_LANES[$i]['type'],
 			$FANNIE_LANES[$i]['op'],$FANNIE_LANES[$i]['user'],
 			$FANNIE_LANES[$i]['pw']);
+        
+        if (!isset($laneupdate_sql->connections[$FANNIE_LANES[$i]['op']]) || $laneupdate_sql->connections[$FANNIE_LANES[$i]['op']] === false) {
+            // connect failed
+            continue;
+        }
 
 		// generate list of columns that exist on both
 		// the server and the lane
@@ -72,6 +77,10 @@ function deleteProductAllLanes($upc){
 		$tmp = new SQLManager($FANNIE_LANES[$i]['host'],$FANNIE_LANES[$i]['type'],
 			$FANNIE_LANES[$i]['op'],$FANNIE_LANES[$i]['user'],
 			$FANNIE_LANES[$i]['pw']);
+        if (!isset($tmp->connections[$FANNIE_LANES[$i]['op']]) || $tmp->connections[$FANNIE_LANES[$i]['op']] === false) {
+            // connect failed
+            continue;
+        }
 		$delQ = $tmp->prepare_statement("DELETE FROM products WHERE upc=?");
 		$delR = $tmp->exec_statement($delQ,array($upc),$FANNIE_LANES[$i]['op']);
 	}
